refactor(py,3d): simplify two-view pose — shared py-utils, doc cleanup
Post-merge simplify pass on the b3a1769 two-view pose commit.

- **Reuse**: extract `unpack_pts` / `unpack_mat3` / `mat3_to_py` / `mask_to_py`
  to a new `kornia-py/src/pyutils.rs`. The functions were duplicated between
  `homography.rs` and `twoview.rs` with identical implementations.
- **Doc rot**: drop the EuRoC-specific numbers from the `use_5pt_essential`
  field doc (numbers belong in `benchmarks.md`, not in stable API docs that
  rot when the bench evolves).
- **Comment quality**: reword the cheirality "stays in one unit" inline
  comment to explain the *why* (predicate consistency for the ambiguity
  ratio) rather than narrate the implementation.

No behavior change. 67 Rust pose tests + 16 Python tests pass; bench shows
identical 3.389° t-err / 2.96 ms pose / 110 matches / 83 inliers.
